+++ ChangeLog.md
@@ -3,7 +3,8 @@ XP Reflection ChangeLog
 
 ## ?.?.? / ????-??-??
 
-* Implemented context type check when invoking methods and constructors
+* Implemented context type check when reading and writing properties as
+  well as when invoking methods and constructors
   (@thekid)
 
 ## 0.2.0 / 2020-12-12